Clear collection optimization
Added the feature I previously mentioned in the readme that takes the number of items available and only clicks that many times. It's much faster now but sometimes throws out some errors. I will have to create a new collection and delete it to test to figure out these errors.